dyninst: don't use the batcher for uploads
Move away from using the Batcher for uploading SymDB. The Batcher was
written for logs, which are pretty different -- they're continuous in
nature and so the main point of the Batcher is to flush on a timer. We
don't need the timer behavior for SymDB upload, and so we were gaining
very little for using it for SymDB; replacing it is only a few lines of
code. Sharing the logic came with a bunch of issues:
- we had multiple concurrency levels, making it more difficult to reason
  about the code: requestUpload was spawning a goroutine and the also
  the batcher was spawning goroutines for each upload batch
- because of the async nature, the Batcher allowed for out of order
  uploads. This will be a problem if we start uploading "amend" batches,
  which rely on some scopes already existing in the database
- the error behavior is funky -- the interface currently doesn't return
  any upload errors (although I had a patch to store the first error and
  return it on Stop())
- (this is the big one) the fact that uploads were async, decoupled from
  enqueuing the data, makes the memory usage unpredictable -- there's no
  control over how many uploads are in flight, which can mean that the
  whole symdb data is effectively in memory at once. Instead, we want slow
  uploads to backpressure the dwarf exploration.

We could make the batcher more configurable to address some of these,
but it's not worth it; the SymDB code is better of not depending on it
and doing its own batching and synchronous uploading.